polymarket-watch/pyproject.toml
[tool.poetry]
name = "polymarket-watch"
version = "0.1.0"
description = "Monitoring scaffold for Polymarket tracking."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "polymarket_watch"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
httpx = "^0.26.0"
python-telegram-bot = "^20.8"
psycopg = {extras = ["binary"], version = "^3.1.18"}
setuptools = "^80.9.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
honcho = "^1.1.0"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "B", "I", "UP", "N", "A", "S", "C4", "Q", "TID", "T10"]
ignore = ["A003"]

[tool.ruff.per-file-ignores]
"db/migrations/**" = ["E402", "F401"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
filterwarnings = [
    "ignore:.*iscoroutinefunction.*:DeprecationWarning:pytest_asyncio.plugin",
    "ignore:.*get_event_loop_policy.*:DeprecationWarning:pytest_asyncio.plugin",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
